pyproject.toml
[project]
name = "GraphragKM"
version = "0.1.1"
description = "GraphRAG 驱动的 AI 本体生成工具"
authors = [{ name = "hanpeilin", email = "alchemylinux@gmail.com" }]
readme = "README.md"
requires-python = ">=3.11,<3.13"
license = { text = "MIT" }
dependencies = [
    "openai (>=1.65.4,<2.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "graphrag (>=2.0.0,<3.0.0)",
    "easyocr (>=1.7.2,<2.0.0)",
    "rdflib (>=7.1.3,<8.0.0)",
    "pathlib (>=1.0.1,<2.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "numpy (>=1.26.4,<2.0.0)",
    "rich (>=13.9.4,<14.0.0)",
    "matplotlib (>=3.10.1,<4.0.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "click (>=8.1.8,<9.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "opencv-python-headless (>=4.11.0.86,<5.0.0.0)",
]

[tool.poetry]
packages = [{ include = "graphragkm", from = "src" }]


[tool.poetry.group.dev.dependencies]
notebook = "^7.3.2"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
graphragkm = "graphragkm.cli:main"